refactor(decomposition): decompose color picker and improve styles
