+++ src/components/UI/Button.tsx
@@ -18,7 +18,7 @@ const buttonVariants = cva(
 				secondary:
 					'bg-secondary text-secondary-foreground shadow-xs hover:bg-secondary/80',
 				ghost:
-					'hover:bg-accent hover:text-accent-foreground dark:hover:bg-accent/50',
+					'hover:bg-accent hover:text-background dark:hover:bg-accent/50',
 				link: 'text-primary underline-offset-4 hover:underline hover:text-accent',
 			},
 			size: {
